feat: Implement SearchProfileEditorDialog for streamlined profile management and display active filters as badges in the FilterPanel.
