+++ apps/web/src/pages/SearchProfilesPage.tsx
@@ -5,14 +5,11 @@ import { toast } from 'sonner'
 import { useTranslation } from 'react-i18next'
 import { rawApiClient } from '@/lib/api-helpers'
 import { ProfileCard, type SearchProfileRunStatus, type SearchProfileSummary } from '@/components/ProfileCard'
-import { JobDescriptionSelect } from '@/components/JobDescriptionSelect'
 import { Button } from '@/components/ui/button'
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
 import { Dialog, DialogContent, DialogFooter, DialogHeader, DialogTitle } from '@/components/ui/dialog'
-import { Input } from '@/components/ui/input'
-import { Label } from '@/components/ui/label'
-import { Checkbox } from '@/components/ui/checkbox'
 import { PageHeader } from '@/components/PageHeader'
+import { SearchProfileEditorDialog } from '@/components/SearchProfileEditorDialog'
 
 type SearchProfileDetails = {
   id: string
@@ -56,32 +53,9 @@ type ProfileStatusResponse = {
   status?: SearchProfileRunStatus | null
 }
 
-type ProfileFormState = {
-  name: string
-  location: string
-  keywordsText: string
-  jobDescription: string
-  cron: string
-  enabled: boolean
-}
-
-const DEFAULT_FORM: ProfileFormState = {
-  name: '',
-  location: '东莞',
-  keywordsText: '',
-  jobDescription: '',
-  cron: '0 9 * * 1-5',
-  enabled: true,
-}
-
 const TERMINAL_STATUSES: Array<SearchProfileRunStatus['taskStatus']> = ['completed', 'failed', 'cancelled', 'unknown']
 
-function parseKeywords(value: string): string[] {
-  return value
-    .split(/[\s,，、]+/)
-    .map((keyword) => keyword.trim())
-    .filter((keyword) => keyword.length > 0)
-}
+
 
 function isRecord(value: unknown): value is Record<string, unknown> {
   return typeof value === 'object' && value !== null
@@ -143,23 +117,11 @@ function buildScheduleLabel(profile?: SearchProfileDetails): string {
   return profile.schedule.cron || 'enabled'
 }
 
-function toFormState(profile: SearchProfileDetails): ProfileFormState {
-  return {
-    name: profile.name,
-    location: profile.location,
-    keywordsText: profile.keywords.join(' '),
-    jobDescription: profile.jobDescription || '',
-    cron: profile.schedule?.cron || '',
-    enabled: profile.status === 'active',
-  }
-}
-
 export function SearchProfilesPage() {
   const { t } = useTranslation()
   const [searchParams, setSearchParams] = useSearchParams()
 
   const [loading, setLoading] = useState(true)
-  const [submitting, setSubmitting] = useState(false)
   const [profiles, setProfiles] = useState<SearchProfileSummary[]>([])
   const [profileDetails, setProfileDetails] = useState<Record<string, SearchProfileDetails>>({})
   const [runStatuses, setRunStatuses] = useState<Record<string, SearchProfileRunStatus>>({})
@@ -167,7 +129,6 @@ export function SearchProfilesPage() {
   const [editorOpen, setEditorOpen] = useState(false)
   const [editingProfileId, setEditingProfileId] = useState<string | null>(null)
   const [deletingProfileId, setDeletingProfileId] = useState<string | null>(null)
-  const [form, setForm] = useState<ProfileFormState>(DEFAULT_FORM)
 
   const pollTimersRef = useRef<Record<string, ReturnType<typeof setInterval>>>({})
 
@@ -274,34 +235,13 @@ export function SearchProfilesPage() {
 
   const handleCreate = useCallback(() => {
     setEditingProfileId(null)
-    setForm(DEFAULT_FORM)
     setEditorOpen(true)
   }, [])
 
   const handleEdit = useCallback(async (profileId: string) => {
-    const cached = profileDetails[profileId]
-    if (cached) {
-      setEditingProfileId(profileId)
-      setForm(toFormState(cached))
-      setEditorOpen(true)
-      return
-    }
-
-    const { data } = await rawApiClient.GET<ProfileResponse>(`/api/search-profiles/${profileId}`)
-    const profile = data?.profile
-    if (!data?.success || !profile) {
-      toast.error(t('searchProfiles.loadDetailError', { defaultValue: 'Failed to load profile details' }))
-      return
-    }
-
-    setProfileDetails((previous) => ({
-      ...previous,
-      [profileId]: profile,
-    }))
     setEditingProfileId(profileId)
-    setForm(toFormState(profile))
     setEditorOpen(true)
-  }, [profileDetails, t])
+  }, [])
 
   const handleDelete = useCallback((profileId: string) => {
     setDeletingProfileId(profileId)
@@ -380,55 +320,7 @@ export function SearchProfilesPage() {
     }
   }, [startPolling, t])
 
-  const handleSave = useCallback(async () => {
-    const keywords = parseKeywords(form.keywordsText)
-    if (!form.name.trim() || !form.location.trim() || keywords.length === 0) {
-      toast.error(t('searchProfiles.validationError', { defaultValue: 'Name, location and keywords are required' }))
-      return
-    }
 
-    const payload = {
-      name: form.name.trim(),
-      location: form.location.trim(),
-      keywords,
-      status: form.enabled ? 'active' : 'paused',
-      jobDescription: form.jobDescription.trim() || undefined,
-      schedule: {
-        enabled: form.enabled,
-        cron: form.cron.trim() || undefined,
-      },
-    }
-
-    setSubmitting(true)
-    try {
-      if (editingProfileId) {
-        const { data } = await rawApiClient.PUT<{ success: boolean }>(`/api/search-profiles/${editingProfileId}`, {
-          body: payload,
-        })
-        if (!data?.success) {
-          throw new Error('Failed to update profile')
-        }
-      } else {
-        const { data } = await rawApiClient.POST<{ success: boolean }>('/api/search-profiles', {
-          body: payload,
-        })
-        if (!data?.success) {
-          throw new Error('Failed to create profile')
-        }
-      }
-
-      setEditorOpen(false)
-      setEditingProfileId(null)
-      setForm(DEFAULT_FORM)
-      toast.success(t('searchProfiles.saveSuccess', { defaultValue: 'Profile saved' }))
-      await loadProfiles()
-    } catch (error) {
-      console.error('Failed to save profile', error)
-      toast.error(t('searchProfiles.saveError', { defaultValue: 'Failed to save profile' }))
-    } finally {
-      setSubmitting(false)
-    }
-  }, [editingProfileId, form, loadProfiles, t])
 
   useEffect(() => {
     const editId = searchParams.get('edit')
@@ -509,85 +401,12 @@ export function SearchProfilesPage() {
         </div>
       )}
 
-      <Dialog open={editorOpen} onOpenChange={setEditorOpen}>
-        <DialogContent className="sm:max-w-xl">
-          <DialogHeader>
-            <DialogTitle>
-              {editingProfileId
-                ? t('searchProfiles.editTitle', { defaultValue: 'Edit Profile' })
-                : t('searchProfiles.createTitle', { defaultValue: 'Create Profile' })}
-            </DialogTitle>
-          </DialogHeader>
-
-          <div className="grid gap-4 py-2">
-            <div className="grid gap-2">
-              <Label htmlFor="profile-name">{t('searchProfiles.fields.name', { defaultValue: 'Name' })}</Label>
-              <Input
-                id="profile-name"
-                value={form.name}
-                onChange={(event) => setForm((previous) => ({ ...previous, name: event.target.value }))}
-              />
-            </div>
-
-            <div className="grid gap-2">
-              <Label htmlFor="profile-location">{t('searchProfiles.fields.location', { defaultValue: 'Location' })}</Label>
-              <Input
-                id="profile-location"
-                value={form.location}
-                onChange={(event) => setForm((previous) => ({ ...previous, location: event.target.value }))}
-              />
-            </div>
-
-            <div className="grid gap-2">
-              <Label htmlFor="profile-keywords">{t('searchProfiles.fields.keywords', { defaultValue: 'Keywords' })}</Label>
-              <Input
-                id="profile-keywords"
-                value={form.keywordsText}
-                onChange={(event) => setForm((previous) => ({ ...previous, keywordsText: event.target.value }))}
-                placeholder={t('searchProfiles.fields.keywordsPlaceholder', { defaultValue: 'e.g. 车床 销售 CNC' })}
-              />
-            </div>
-
-            <div className="grid gap-2">
-              <Label>{t('searchProfiles.fields.jobDescription', { defaultValue: 'Job Description' })}</Label>
-              <JobDescriptionSelect
-                value={form.jobDescription}
-                onChange={(value) => setForm((previous) => ({ ...previous, jobDescription: value }))}
-              />
-            </div>
-
-            <div className="grid gap-2">
-              <Label htmlFor="profile-cron">{t('searchProfiles.fields.cron', { defaultValue: 'Cron Expression' })}</Label>
-              <Input
-                id="profile-cron"
-                value={form.cron}
-                onChange={(event) => setForm((previous) => ({ ...previous, cron: event.target.value }))}
-                placeholder="0 9 * * 1-5"
-              />
-            </div>
-
-            <div className="flex items-center gap-2">
-              <Checkbox
-                checked={form.enabled}
-                onCheckedChange={(checked) => setForm((previous) => ({ ...previous, enabled: checked === true }))}
-                id="profile-enabled"
-              />
-              <Label htmlFor="profile-enabled">{t('searchProfiles.fields.enabled', { defaultValue: 'Enabled' })}</Label>
-            </div>
-          </div>
-
-          <DialogFooter>
-            <Button variant="outline" onClick={() => setEditorOpen(false)}>
-              {t('searchProfiles.cancel', { defaultValue: 'Cancel' })}
-            </Button>
-            <Button onClick={() => void handleSave()} disabled={submitting}>
-              {submitting
-                ? t('searchProfiles.saving', { defaultValue: 'Saving...' })
-                : t('searchProfiles.save', { defaultValue: 'Save' })}
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
+      <SearchProfileEditorDialog
+        open={editorOpen}
+        onOpenChange={setEditorOpen}
+        profileId={editingProfileId}
+        onSaved={loadProfiles}
+      />
 
       <Dialog open={!!deletingProfileId} onOpenChange={(open) => !open && setDeletingProfileId(null)}>
         <DialogContent className="sm:max-w-md">
